pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vertice-code"
version = "1.0.0"
description = "Vértice: Sovereign Intelligence & Tactical Execution - Enterprise Multi-LLM Orchestration"
readme = "README.md"
authors = [
    {name = "Juan Carlos de Souza", email = "juancs.dev@gmail.com"}
]
license = "MIT"
requires-python = ">=3.11"
keywords = ["ai", "code-assistant", "mcp", "llm", "multi-model", "vertice", "orchestration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core CLI/TUI
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "pyfiglet>=1.0.0",
    "pyperclip>=1.8.0",
    "textual>=0.89.0",
    "pydantic>=2.0.0",

    # Resilience (audit fix: tenacity missing)
    "tenacity>=8.2.0",
    "cryptography>=42.0.0", # SEC-004: Fernet encryption for vault
    "aiohttp>=3.9.0",

    # Graph Analysis
    "networkx>=3.0",

    # LLM Providers
    "google-cloud-aiplatform>=1.74.0",  # Vertex AI with Gemini 2.0
    "google-generativeai>=0.8.0",        # AI Studio fallback
    "anthropic>=0.40.0",                 # Claude API
    "openai>=1.50.0",                    # OpenAI/Azure
    "groq>=0.11.0",                      # Groq (free tier)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
fast = [
    "uvloop>=0.18.0",
]
# Additional LLM providers (optional)
providers = [
    "mistralai>=0.1.0",
    "cohere>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/JuanCS-Dev/vertice-code"
Repository = "https://github.com/JuanCS-Dev/vertice-code"
Issues = "https://github.com/JuanCS-Dev/vertice-code/issues"

[project.scripts]
# TUI (default - main interface)
vertice = "vertice_cli.main:cli_main"

# CLI entry points (alternative)
vtc = "vertice_cli.main:cli_main"
vertice-cli = "vertice_cli.main:cli_main"

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "dist",
    "build",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
include = ["vertice_cli*", "vertice_tui*", "vertice_core*", "vertice_governance*", "prometheus*"]
exclude = ["vertice_core.onboarding", "vertice_core.pilot_management", "vertice_core.core"]
namespaces = false
